Conflicts:
	benchmarks/data/disk.cpp
	c7a/api/context.cpp
	c7a/api/context.hpp
	c7a/api/generate.hpp
	c7a/api/input_line_iterator.hpp
	c7a/api/read.hpp
	c7a/api/reduce_to_index.hpp
	c7a/common/atomic_movable.hpp
	c7a/data/channel.hpp
	c7a/data/channel_multiplexer.hpp
	c7a/data/channel_sink.hpp
	c7a/net/construction.hpp
	c7a/net/dispatcher.hpp
	c7a/net/dispatcher_thread.hpp
	tests/api/operations_test.cpp

+++ c7a/core/reduce_pre_probing_table.hpp
@@ -251,7 +251,7 @@ public:
         KeyValuePair* initial = &vector_[h.global_index];
         KeyValuePair* current = initial;
         KeyValuePair* next_partition = &vector_[h.global_index -
-                (h.global_index % num_items_per_partition_) + num_items_per_partition_];
+                                                (h.global_index % num_items_per_partition_) + num_items_per_partition_];
 
         while (!equal_to_function_(current->first, sentinel_.first))
         {
@@ -300,7 +300,8 @@ public:
             FlushLargestPartition();
         }
 
-        if ((float)items_per_partition_[h.partition_id] / (float)num_items_per_partition_
+        if (static_cast<double>(items_per_partition_[h.partition_id]) /
+            static_cast<double>(num_items_per_partition_)
             > max_partition_fill_ratio_)
         {
             LOG << "resize";
